// src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Date: { input: any; output: any; }
  JSON: { input: any; output: any; }
};
// ...
export enum IdCode {
  Aadhar = 'Aadhar',
  Pan = 'PAN'
}
// ...
export enum ModeOfPayment {
  Cash = 'cash',
  Cheque = 'cheque',
  Online = 'online'
}
// ...
export type Receipt = {
  __typename?: 'Receipt';
  aadharNumber?: Maybe<Scalars['String']['output']>;
  address?: Maybe<Scalars['String']['output']>;
  amount: Scalars['Int']['output'];
  cancelled?: Maybe<Scalars['Boolean']['output']>;
  date?: Maybe<Scalars['Date']['output']>;
  financialYear?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  idCode?: Maybe<IdCode>;
  mobileNumber?: Maybe<Scalars['String']['output']>;
  modeOfPayment: ModeOfPayment;
  name?: Maybe<Scalars['String']['output']>;
  panNumber?: Maybe<Scalars['String']['output']>;
  receiptBook: ReceiptBook;
  receiptNumber: Scalars['Int']['output'];
  uuid: Scalars['String']['output'];
};

export type ReceiptBook = {
  __typename?: 'ReceiptBook';
  financialYear?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  receiptBookNumber: Scalars['Int']['output'];
  receiptSeries: Scalars['Int']['output'];
  receipts?: Maybe<Array<Maybe<Receipt>>>;
  totalReceipts: Scalars['Int']['output'];
  usedReceipts: Scalars['Int']['output'];
  uuid: Scalars['String']['output'];
};